Decompose tmux session file
