+++ docs/file-length-lint.md
@@ -39,7 +39,6 @@ splitting them):
 | File                       | Ceiling |
 | -------------------------- | ------- |
 | `session/instance.go`      | 1400    |
-| `session/tmux/tmux.go`     | 1393    |
 | `ui/sidebar.go`            | 1258    |
 | `config/config.go`         | 1128    |
 | `session/backend_hook.go`  | 1026    |
