+++ scripts/file-length-allowlist.txt
@@ -18,7 +18,6 @@
 
 # --- production code (limit 1000) ---
 session/instance.go 1400
-session/tmux/tmux.go 1393
 ui/sidebar.go 1258
 config/config.go 1128
 session/backend_hook.go 1026
